moved counter to different file in components
