+++ README.md
@@ -26,3 +26,14 @@ All scripts can be run from the root.
 
 - Node
 - Yarn (can be installed using `npm install -g yarn`)
+
+## Cors
+We have cors within the backend disabled right now, but we need to enable it when deploying. It could be a secure risk.
+```
+const corsOptions ={
+  origin:'http://localhost:4000', 
+  credentials:true,            //access-control-allow-credentials:true
+  optionSuccessStatus:200
+}
+
+```
