+++ packages/backend/server/index.ts
@@ -1,16 +1,17 @@
 const express = require("express");
 
 const morgan = require('morgan')
-const cors = require('cors')
-const corsOptions ={
-  origin:'http://localhost:4000', 
-  credentials:true,            //access-control-allow-credentials:true
-  optionSuccessStatus:200
-}
+// const cors = require('cors')
+
+// const corsOptions ={
+//   origin:'http://localhost:4000', 
+//   credentials:true,            //access-control-allow-credentials:true
+//   optionSuccessStatus:200
+// }
 
 
 const app = express(); 
-app.use(cors(corsOptions))
+// app.use(cors(corsOptions))
 app.use(morgan('dev'))
 
 app.use(express.json())
