+++ packages/frontend/webpack.config.js
@@ -35,9 +35,13 @@ module.exports = {
     filename: "bundle.js"
   },
   devServer: {
-    port: 4000,
+    port: 3000,
     open: true,
-    hot: true
+    hot: true,
+    proxy: {
+      '/api': 'https://localhost:5000',
+      pathRewrite: { '^/api': '' },
+    },
   },
   plugins: [
     new HtmlWebpackPlugin({
